requirements.txt
# --- Core OpenAI & LangChain stack ---
openai>=1.44.0
langchain>=0.3.2
langchain-core>=0.3.5
langchain-community>=0.3.2
langchain-openai>=0.3.0
langchain-cerebras>=0.2.0
langchain-huggingface>=0.1.0

# --- Embeddings & ML ---
sentence-transformers>=3.0.1
transformers>=4.44.2
torch>=2.4.1
torchvision>=0.19.1
faiss-cpu>=1.8.0

# --- Vector DBs & tools ---
chromadb>=0.5.5

# --- Document loaders / processing ---
pypdf>=4.2.0
unstructured>=0.15.0

# --- Utilities ---
python-dotenv>=1.0.1
guardrails-ai>=0.4.1
smolagents>=0.1.4
wikipedia>=1.4.0
google-genai>=0.4.0
ddgs>=2.3.0
langchain-mcp


# --- UI / Interaction ---
gradio>=4.44.0
playwright>=1.48.0
